refactor: перенести общие для проекта классы в пакет common.
