+++ src/main/java/ru/practicum/shareit/common/ErrorHandler.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit;
+package ru.practicum.shareit.common;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
