+++ src/main/java/ru/practicum/shareit/common/ErrorResponse.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit;
+package ru.practicum.shareit.common;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
