+++ src/main/java/ru/practicum/shareit/common/Update.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.common;
+
+public interface Update {
+}
