+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import ru.practicum.shareit.common.Create;
 import ru.practicum.shareit.requests.dto.ItemRequestDto;
 
 import javax.validation.constraints.NotBlank;
@@ -17,13 +18,13 @@ public class ItemDto {
 
     private Long id;
 
-    @NotBlank(groups = ItemCreateMarker.class, message = "Не указано название вещи.")
+    @NotBlank(groups = Create.class, message = "Не указано название вещи.")
     private String name;
 
-    @NotBlank(groups = ItemCreateMarker.class, message = "Не указано описание вещи.")
+    @NotBlank(groups = Create.class, message = "Не указано описание вещи.")
     private String description;
 
-    @NotNull(groups = ItemCreateMarker.class, message = "Не указана доступность вещи.")
+    @NotNull(groups = Create.class, message = "Не указана доступность вещи.")
     private Boolean available;
 
     private ItemRequestDto request;
