--- src/main/java/ru/practicum/shareit/item/dto/ItemUpdateMarker.java
@@ -1,4 +1,0 @@
-package ru.practicum.shareit.item.dto;
-
-public interface ItemUpdateMarker {
-}
