--- src/main/java/ru/practicum/shareit/user/dto/UserCreateMarker.java
@@ -1,4 +1,0 @@
-package ru.practicum.shareit.user.dto;
-
-public interface UserCreateMarker {
-}
