--- src/main/java/ru/practicum/shareit/user/dto/UserUpdateMarker.java
@@ -1,4 +1,0 @@
-package ru.practicum.shareit.user.dto;
-
-public interface UserUpdateMarker {
-}
